Changes to work with ModularShaderLanguage.
Added submodule for SPIRV-Headers. This simplifies the checkout process and
ensures a consistent revision. Otherwise, a manual checkout can easily have
incompatibilities, such as adding new enums and warning about unused
values, causing the build to spontaneously fail.

Deleted generation of the shared library build, since this will never be
used. Commented out inclusion of all the sub-libraries except for opt,
since that is the only one that's currently needed.